refact: Move initWidgets function to execute once
Since there are 2 extended instances of
HsQueryPopupBaseService (OL and Cesium), initWidgets gets executed twice
resulting in doubled widget list.
